update bocadillo app file structure, use 0.14-compatible APIs
