+++ server/asgi.py
@@ -1,0 +1,5 @@
+from bocadillo import configure
+from .app import app
+from . import settings
+
+configure(app, settings)
